pyproject.toml
[tool.poetry]
name = "asr_service"
version = "0.1.0"
description = ""
authors = ["TJ"]
packages = [
    { include = "asr_service", from = "src" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "2.1.0"
huggingface = "^0.0.1"
transformers = "^4.36.0"
accelerate = "0.26.0"
datasets = {extras = ["audio"], version = "^2.16.1"}
librosa = "^0.10.1"
torchaudio = "^2.1.2"
pydub = "0.25.1"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
